+++ js/form-status.js
@@ -6,7 +6,6 @@ const mapFeature = mapForm.querySelector('.map__features');
 const featuresInput = mapFeature.querySelectorAll('.map__checkbox');
 
 const setFormInactive = () => {
-  console.log('jjj');
   mainForm.classList.add('ad-form--disabled');
 
   for (let i = 0; i < formFieldsets.length; i++) {
@@ -24,8 +23,9 @@ const setFormInactive = () => {
   }
 };
 
+setFormInactive();
+
 const setFormActive = () => {
-  console.log('aaa');
   mainForm.classList.remove('ad-form--disabled');
   for (let i = 0; i < formFieldsets.length; i++) {
     formFieldsets[i].disabled = false;
