+++ js/form-validation.js
@@ -1,7 +1,7 @@
 import { sendData } from './api.js';
 import { openErrorMessage} from './message.js';
-import { getFormInitialState, checkboxReset, filterReset } from './form.js';
-import {clearLayers} from './map.js';
+import { getMapInitialState, clearLayers } from './map.js';
+import { openSuccessMessage } from './message.js';
 
 const form = document.querySelector('.ad-form');
 const validRooms = document.querySelector('#room_number');
@@ -13,6 +13,7 @@ const validTimeOut = document.querySelector('#timeout');
 const mapFilter = document.querySelector('.map__filters');
 const filterFeatures = mapFilter.querySelectorAll('[name="features"]');
 const mapFeatures = mapFilter.querySelectorAll('select');
+const buttonReset = document.querySelector('.ad-form__reset');
 
 
 const pristine = new Pristine(form, {
@@ -111,6 +112,37 @@ validTimeOut.addEventListener('change', () => {
   getValdTimeOut();
 });
 
+
+const checkboxReset = (element) => {
+  element.forEach((value) => {
+    value.checked = false;
+  });
+};
+
+const filterReset = (element) => {
+  element.forEach((el) => {
+    el.value = 'any';
+  });
+};
+
+//сброс по кнопке 'очистить'
+buttonReset.addEventListener('click', (evt) => {
+  evt.preventDefault();
+  form.reset();
+  getMapInitialState();
+  clearLayers();
+  filterReset(mapFeatures);
+  checkboxReset(filterFeatures);
+});
+
+
+const getFormInitialState = () => {
+  openSuccessMessage();
+  form.reset();
+  getMapInitialState();
+};
+
+
 const setUserFormSubmit = () => {
   form.addEventListener('submit', (evt) => {
     evt.preventDefault();
@@ -126,3 +158,4 @@ const setUserFormSubmit = () => {
 
 setUserFormSubmit();
 
+export {getFormInitialState, checkboxReset, filterReset};
